Type conversion (float -> int) in casingress.py altitude handling.

+++ game/missiongenerator/aircraft/waypoints/casingress.py
@@ -68,6 +68,6 @@ class CasIngressBuilder(PydcsWaypointBuilder):
             Bf_109K_4,
             I_16,
         ]:
-            waypoint.alt = waypoint.alt / 3
+            waypoint.alt = int(waypoint.alt / 3)
         else:
-            waypoint.alt = waypoint.alt / 2
+            waypoint.alt = int(waypoint.alt / 2)
